+++ common/src/main/java/com/kryeit/missions/MissionManager.java
@@ -183,6 +183,7 @@ public class MissionManager {
         ServerPlayer serverPlayer = playerList.getPlayer(player);
         if (serverPlayer == null) return;
 
+        Utils.increaseMissionStat(serverPlayer, mission);
         serverPlayer.connection.send(new ClientboundSoundPacket(Holder.direct(ModSounds.MISSION_COMPLETE.get()), SoundSource.NEUTRAL, serverPlayer.position().x, serverPlayer.position().y, serverPlayer.position().z, 1, 1, 1));
         Utils.executeCommandAsServer(COMMAND_UPON_MISSION.replace("%player%", serverPlayer.getName().getString()));