+++ common/src/main/java/com/kryeit/registry/ModTags.java
@@ -6,6 +6,7 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
+
 public class ModTags {
 
     // Common Tags