+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.3.3
 - Fix #32: Certain modded items not rendering in Reroll Button
 - Added Relocate Mission type
+- Added Vanilla stats for: Easy, normal and hard total completed missions
 
 # 0.3.2
 - Fixed some dedicated servers not saving properly the mission data